requirements.txt
selenium==4.25.0
pytest==8.3.3
allure-pytest==2.13.5
allure-python-commons~=2.13.5